Some LoRa fixes
The port number is now 10 (to prevent casual experiments on lower ports
to interfere). No trailing zeroes are transmitted, since the receiver
can just look at the packet length to see what is present. The datarate
is fixed to SF9, since ADR probably does not work yet and this is the
most reliable DR that still allows sending packets every 15minutes
within the TTN fair usage policy. Finally, do not check for TXRXPEND,
just overwrite a previous packet if it is present.