Rewrite the sketch for the heatmap slam
With these changes, the GPS is kept on continuously and everytime at
least 15 seconds pass and 50 meter is moved, a measurement is made. Up
to 10 measurements are stored, and whenever a new measurement is made
and airtime is available, all stored measurements are transmitted.

To reduce packet size, one full measurement is sent, plus incremental
differences. To allow for big and small changes, the number of bits used
for these differences is also transmitted as part of the packet.

The supply voltage is only transmitted once, along with the most recent
measurement since it is not as important.