Remove pointless typecasts
When encoding the temperature and humidity, the float value was first
cast to uint16_t and then assigned to an int16_t variable. This is
remnant of older code, where a bitshift or mask was applied before
assignassignment, which does not work on a float, so the cast was
needed. This bitfiddling is now handled by BitStream, so the cast is not
needed.

Note that the cast used uin16_t, which suggests that negative
temperatures were not handled properly. However, casting a negative
float to an unsigned type is implementation-defined and it seems avr-gcc
does not truncat to 0, but instead underflows, resulting in the proper
value when the uint16_t was next cast to int16_t, so negative values
already worked.