Send a firmware version number in the packet
This version number is stored in a constant in the code and should be
manually incremented when significant changes happen and/or a lot of
nodes are upgraded.

Since this changes the format of the packet (and just checking the
packet length is ambiguous, since the battery voltage byte is already
optional), the port number is incremented.