Retry sending data when no airtime is available
Previously, when no airtime is available (or joining is not complete),
the message would not be resent until a new measurement is added (e.g.
when enough time and distance have passed). Now, the message is retried
until airtime is available.

In particular, this means that the first message (which fails because
joining is not complete) is actually sent after the join completes,
without having to cross some distance as was previously the case.