Change sketch into simple rangetesting app
Instead of collecting multiple position, temperature and humidity
readings and sending a history of them, this now just sends a single GPS
reading. It also adds the altitude and HDOP (accuracy) information to
the packet. This is intended as a range testing app, to send data to
ttnmapper.org.

To disambiguate from the previous sketch, this uses a different LoRaWAN
port (21).