Re-initialize GPS and GPS serial on every attempt
The GPS and GPS serial objects are now no longer global, but are
allocated inside `getPosition()` when needed. This ensures that they are
always cleanly initialized and additionally saves some global memory
(which is now only allocated when it is actually being used).

This change seems to improve the stability of GPS readouts (either this
commit, or the previous one does).