Allow extra fields in the packet
In the future, we want to switch to a much more flexible packet format
that allows arbitrary fields. However, to have something that can be
for experiments in the meanwhile, this allows specifying extra integer
fields of arbitrary size.

Previously, the packet length was used to determine the presence of some
optional field (and the port number for the presence of the lux field,
which would otherwise be ambiguous). When an arbitrary number of bytes
can be appended at the end, using the packet length for this will always
be ambiguous. To prevent that, an extra byte is inserted at the start of
the packet containing some flags for some of the optional fields.

To distinguish between this new packet format and the older ones, a new
port number is used.

By default, no extra fields are sent, but some (commented) code is
present as an example.