Allow lux values 4 times higher
This scales the lux value by 4 (2 bits) when transmitting, to allow up
to 256k lux in a 16-bit value. This is not currently used, since the
resistor values limit to something like 40k, but to allow expanding this
in the future, the packet format for port 13 was defined with 2-bit
scaling.

The current maximum of 40k has already proven to be insufficient for
sunny days. The full solar power should be around 128k, so 256k should
be sufficient.

This additionally enlarges the `lux` variable to 32-bits to store the
larger value (saving the scaling for when the packet is being built).
The `readLux` function already returned a uint32_t.