Use explicitly-sized integer types
This replaces `long` by `uint32_t`, which on AVR are the same size, but
using uint32_t is more explicit and portable.

This also replaces some `int` with `uint8_t`, which is more than big
enough for what they are used for.

Note that some timer values still use `unsigned long`, but this is how
the Arduino API is defined, so these are intentionally left like that.