Experiment with two luxmeters
This code reads out two luxmeters and puts their calculated lux values,
raw adc value and range in the extra data. The calculated lux value of
the first meter is also put into the normal lux field.

This code is written for an experiment comparing two covers for the
luxmeter (pingpongball and 3mm teflon) to be mounted near an official
light sensor in Bergen. The R11 value is lowered from 10k to 3.3k for
some additional range.