Fix handling of packet overflow
In the BitStream class, if one would try to add more data than there is
capacity, previous data would erronously be removed from the buffer.
This is probably a remnant of an earlier implementation where the
pos pointer would be incremented before checking for overflow, so it
would need to be retracted on error.

Note that this condition would never occur in practice, since the
maximum amount of data added to the buffer is fixed and the buffer is
sized appropriately.