Do a LoRaWAN join before starting the loop
Previously, the OTAA join would be automatic, as soon as the first
packet is queued. However, this gives only little control over the join,
ends up sending data that is already at least 5 seconds old and sometimes a
lot more and it seems that when not setting up any channels manually
(next commit), at least BasicMAC (possibly also lmic) does not like
setting a datarate before join started, and BasicMAC (not lmic) does not
like queuing a packet when join is ongoing.

So now the OTAA join is manually started from setup and first completed
explicitly before starting the loop.

This does mean that the GPS is not enabled until after the join, and
that no sleeping happens at all while joining (so with no network
coverage, this drains the battery even more quickly), so this should be
revised later.