Move NMEAGPS object back to global scope
It turns out this object does not explicitly initialize all of its
members. If used as a global variable, they are initialized to zero, but
if used as a local variable, some members remain unitialized, taking
whatever value happens to be in memory.

This was observed to cause a buffer overflow when _firstFix was not
zero, leading to a fault because the htu object was overwritten and then
a reset.

Presumably this did not show up as a problem so far because the object
would be allocated in memory that usually or always contained values
that were ok, or maybe because AVR is less sensitive to e.g. null
pointer dereferencing (no hard fault and reset, but just reading random
values).

This is really a problem in the library, but we can work around it
easily by putting the variable back in global scope.

This problem was introduced in commit fd785f9 (Re-initialize GPS and GPS
serial on every attempt).