Use the LED for status info on MJS2020
This starts out red while waiting for USB/Serial, becomes orange when
doing initial sensor readings, yellow while waiting for TTN activation.
While running or sleeping, the led is off, except when waiting for GPS
(cyan), the SPS030 (purple) or the LoRa transmission (blue).

This adds a bit of power usage, but the led is significantly dimmed, so
it should be up to 100μA or so, which is insignificant compared to the
GPS and SPS030 power usage.