Make sketch work with MJS2020-PROTO3
Note that this reduces the RGB LED PWM by a factor 4, since the led
resistors were are changed from 1k to 3k3, so this keeps about the same
brightness (a lower brightness makes the colors harder to distinguish).
The same scaling is applied for PROTO2, so that will have 4x brighter
leds (and a bit more power usage due to that), but that's ok in order to
keep the code simpler.