Compress extra fields when possible
When transmitting smaller values, use a smaller field size (i.e. do not
bother transmitting leading zeroes).

Because this might end up not using all of the allocated buffer space,
the buffer stuffing at the end of the packet is also slightly changed to
only stuff the current partial byte, rather than the rest of the buffer
(which would fail if there was more than 32 bits left).