Ask for credentials via serial when missing from EEPROM
This allows assigning credentials (application id, device id and key)
to a device after it is assembled, instead of pre-assigning them to
a chip early.

When credentials are entered via serial, they are first tried, and only
after a succesful join they are finalized in EEPROM.

This commit only adds support for the original MJS devices that use
EEPROM, for the MJS2020 devices some additional work is needed to save
to flash (or swith to EEPROM).