Ask for credentials from serial for MJS2020 too
This changes the code to let MJS2020 load credentials from EEPROM if
available, falling back to flash otherwise. If neither contain valid
credentials, credentials can be entered via serial.

There was already code for reading credentials from EEPROM, but it was
controlled via a compiletime switch and we never actually used this in
practice (there should be no boards in the field that have their
credentials in EEPROM). The old code used the same complex layout for
EEPROM as is used for flash, but this commit simplifies that format to
be more similar to the EEPROM format used on AVR.