On MJS2020 write option bytes along with credentials
On older boards, this was done when preconfiguring credentials. Now that
boards no longer need to be preprogrammed, it makes sense to write the
option bytes from within the sketch. This is done at the same time as
saving the credentials, which should happen as one of the first things.

Note thate the board will run fine with the default option bytes too,
just without the brown-out detection enabled.